refactor proxy to eliminate inner class use
